device侧C化应用
# message auto-generated for no-merge-commit merge:
merge main into master

device侧C化应用

Created-by: msprofDeveloper

Match-id-535d3348d4ce352c43243eb4c48ab571edc9df26

+++ analysis/csrc/domain/services/association/ascend_task_association.cpp
@@ -30,7 +30,7 @@ namespace Domain {
 using namespace Analysis::Entities;
 using namespace Utils;
 namespace {
-const uint64_t DEFAULT_MODEL_ID = UINT64_MAX;
+const uint64_t DEFAULT_MODEL_ID = UINT32_MAX;
 const int32_t DEFAULT_INDEX_ID = -1;
 const int64_t DEFAULT_CONNECTION_ID = -1;
 const std::unordered_map<uint32_t, std::string> deviceTaskAcsqTypeMap {
